% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs_edger.R
\name{ngs_edger}
\alias{ngs_edger}
\title{NGS differential expression with edgeR}
\usage{
ngs_edger(counts.file, sample.ids, condition.factor, output.basename,
  output.dir, cutoff.fdr = 0.05, cutoff.lfc = 1, title = "")
}
\arguments{
\item{cutoff.lfc}{}
}
\description{
NGS differential expression with edgeR
}
