% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs_quality_control.R
\name{ngs_fastqc}
\alias{ngs_fastqc}
\title{NGS FastQC}
\usage{
ngs_fastqc(sample.pathnames, output.dir, sample.ids = NULL,
  fastqc.command = "fastqc", multiqc.command = "multiqc", MAX.THREADS = 8,
  title = "")
}
\arguments{
\item{sample.pathnames}{samples to process.}

\item{output.dir}{output directory for results.}

\item{sample.ids}{optional names for the reports.}

\item{fastqc.command}{command to run fastqc.}

\item{MAX.THREADS}{maximum number of concurrent threads to use.}

\item{title}{title of the section}
}
\value{
A list containing the following elements:

chunks: fig1, tab1 and file1.
}
\description{
This function performs quality control of NGS datasets using FASTQC and multiqc. FastQC is run on each sample...
}
