% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs_featureCounts.R
\name{ngs_feature_counts}
\alias{ngs_feature_counts}
\title{NGS counting with featureCounts.}
\usage{
ngs_feature_counts(bam.pathnames, sample.ids, feature.counts.command,
  gene.annotation, output.dir, feature.counts.options = "-t exon -g gene_id",
  MAX.THREADS = 8, title = "")
}
\arguments{
\item{title}{}
}
\description{
NGS counting with featureCounts.
}
