% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs_hisat2.R
\name{ngs_hisat2}
\alias{ngs_hisat2}
\title{NGS mapping with Hisat2}
\usage{
ngs_hisat2(reads1.pathnames, sample.ids, hisat2.command, hisat2.build.command,
  genome.fasta, genome.index.name, output.dir, MAX.THREADS = 8,
  reads2.pathnames = rep(NA, length(reads1.pathnames)), title = "")
}
\arguments{
\item{reads1.pathnames}{samples to map}

\item{sample.ids}{basenames of ouput files}

\item{hisat2.command}{command to run hisat2}

\item{hisat2.build.command}{command to run hisat2-build}

\item{genome.fasta}{genome fasta reference}

\item{genome.index.name}{genome hisat2 index basename}

\item{output.dir}{directory to store output files}

\item{MAX.THREADS}{maximum number of threads to use}

\item{reads2.pathnames}{}
}
\description{
NGS mapping with Hisat2
}
