% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs_quality_control.R
\name{ngs_multiqc}
\alias{ngs_multiqc}
\title{NGS multiqc}
\usage{
ngs_multiqc(input.dir, output.dir, multiqc.command = "multiqc",
  log.basename = NULL, overwrite = FALSE)
}
\arguments{
\item{overwrite}{}
}
\description{
NGS multiqc
}
