% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs_quality_control.R
\name{ngs_quality_control}
\alias{ngs_quality_control}
\title{NGS Quality Control}
\usage{
ngs_quality_control(sample.pathnames, fastqc.command, output.dir,
  sample.ids = NULL, MAX.THREADS = 8, title = "")
}
\arguments{
\item{sample.pathnames}{samples to process.}

\item{fastqc.command}{command to run fastqc.}

\item{output.dir}{output directory for results.}

\item{sample.ids}{optional names for the reports.}

\item{MAX.THREADS}{maximum number of concurrent threads to use.}

\item{title}{title of the section}
}
\value{
A list containing the following elements:

chunks: fig1, tab1 and file1.
}
\description{
This function performs quality control of NGS datasets using FASTQC and multiqc. FastQC is run on each sample...
}
