% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs_trimming.R
\name{ngs_trimming}
\alias{ngs_trimming}
\title{NGS Trimming}
\usage{
ngs_trimming(sample.pathnames, cutadapt.options, cutadapt.command, output.dir,
  sample.ids = NULL, MAX.THREADS = 8)
}
\arguments{
\item{sample.pathnames}{samples to process.}

\item{cutadapt.options}{options to pass to cutadapt command line.}

\item{cutadapt.command}{command to run cutadapt.}

\item{output.dir}{directory to store processed samples.}

\item{sample.ids}{basenames to use for output files.}

\item{MAX.THREADS}{maximum number of concurrent threads to run.}
}
\description{
Performs trimming of fastq files.
}
