% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UBIquitousNGS.R
\docType{package}
\name{UBIquitousNGS-package}
\alias{UBIquitousNGS-package}
\title{UBIquitousNGS: NGS analysis functions for UBIquitous.}
\description{
This package provides functions to make common NGS analyses and report
the results using the UBIquitous framework.
}
\author{
Daniel Neves, \email{danielneves2005@gmail.com}
}
